chore(create-remix): combine installDeps and hasInitScript check

+++ packages/remix-dev/cli/commands.ts
@@ -46,16 +46,14 @@ export async function create({
 
   let initScriptDir = path.join(projectDir, "remix.init");
   let hasInitScript = await fse.pathExists(initScriptDir);
-  if (hasInitScript) {
-    if (installDeps) {
-      console.log("💿 Running remix.init script");
-      await init(projectDir);
-      await fse.remove(initScriptDir);
-    } else {
-      console.log(
-        "💿 You've opted out of installing dependencies so we won't run the remix.init/index.js script for you just yet. Once you've installed dependencies, you can run it manually with `npx remix init`"
-      );
-    }
+  if (installDeps && hasInitScript) {
+    console.log("💿 Running remix.init script");
+    await init(projectDir);
+    await fse.remove(initScriptDir);
+  } else {
+    console.log(
+      "💿 You've opted out of installing dependencies so we won't run the remix.init/index.js script for you just yet. Once you've installed dependencies, you can run it manually with `npx remix init`"
+    );
   }
 
   let relProjectDir = path.relative(process.cwd(), projectDir);
